+++ game_patch/graphics/gr_color.h
@@ -8,15 +8,16 @@
 inline int GetPixelFormatSize(rf::BmPixelFormat pixel_fmt)
 {
     switch (pixel_fmt) {
-    case rf::BMPF_MONO8:
+    case rf::BMPF_BGR_888_INDEXED:
+    case rf::BMPF_A_8:
         return 1;
-    case rf::BMPF_565:
-    case rf::BMPF_4444:
-    case rf::BMPF_1555:
+    case rf::BMPF_RGB_565:
+    case rf::BMPF_RGBA_4444:
+    case rf::BMPF_RGBA_1555:
         return 2;
-    case rf::BMPF_888:
+    case rf::BMPF_RGB_888:
         return 3;
-    case rf::BMPF_8888:
+    case rf::BMPF_RGBA_8888:
         return 4;
     default:
         xlog::warn("unknown pixel format %d", pixel_fmt);
@@ -27,21 +28,26 @@ inline int GetPixelFormatSize(rf::BmPixelFormat pixel_fmt)
 inline D3DFORMAT GetD3DFormatFromPixelFormat(rf::BmPixelFormat pixel_fmt)
 {
     switch (pixel_fmt) {
-    case rf::BMPF_8888:
+    case rf::BMPF_BGR_888_INDEXED:
+        // Note: this indexed format is not really A8 but because there is no corresponding D3D format A8 format
+        // is returned instead (it is used only for pixel size related calculations)
+    case rf::BMPF_A_8:
+        return D3DFMT_A8;
+    case rf::BMPF_RGBA_8888:
         return D3DFMT_A8R8G8B8;
-    case rf::BMPF_888:
+    case rf::BMPF_RGB_888:
         return D3DFMT_R8G8B8;
-    case rf::BMPF_565:
+    case rf::BMPF_RGB_565:
         return D3DFMT_R5G6B5;
-    case rf::BMPF_1555:
+    case rf::BMPF_RGBA_1555:
         return D3DFMT_A1R5G5B5;
-    case rf::BMPF_4444:
+    case rf::BMPF_RGBA_4444:
         return D3DFMT_A4R4G4B4;
     default:
         if (static_cast<unsigned>(pixel_fmt) >= 0x10) {
             return static_cast<D3DFORMAT>(pixel_fmt);
         }
-        xlog::error("unknown pixel format %d", pixel_fmt);
+        xlog::warn("unknown pixel format %d", pixel_fmt);
         return D3DFMT_UNKNOWN;
     }
 }
@@ -50,18 +56,20 @@ inline rf::BmPixelFormat GetPixelFormatFromD3DFormat(D3DFORMAT d3d_fmt)
 {
     switch (d3d_fmt) {
     case D3DFMT_R5G6B5:
-        return rf::BMPF_565;
+        return rf::BMPF_RGB_565;
     case D3DFMT_X1R5G5B5:
     case D3DFMT_A1R5G5B5:
-        return rf::BMPF_1555;
+        return rf::BMPF_RGBA_1555;
     case D3DFMT_A4R4G4B4:
     case D3DFMT_X4R4G4B4:
-        return rf::BMPF_4444;
+        return rf::BMPF_RGBA_4444;
     case D3DFMT_R8G8B8:
-        return rf::BMPF_888;
+        return rf::BMPF_RGB_888;
     case D3DFMT_A8R8G8B8:
     case D3DFMT_X8R8G8B8:
-        return rf::BMPF_8888;
+        return rf::BMPF_RGBA_8888;
+    case D3DFMT_A8:
+        return rf::BMPF_A_8;
     default:
         xlog::error("unknown D3D format 0x%X", d3d_fmt);
         return static_cast<rf::BmPixelFormat>(d3d_fmt);