+++ CHANGELOG.md
@@ -72,6 +72,7 @@ Version 1.4.2
 - fix team scores in TDM
 - fix auto-detection of Steam based Red Faction installation directory
 - fix crash if camera cannot be restored to first-person mode after cutscene (RF bug)
+- fix handling of alpha-only (A8) texture format when 32-bit textures are used
 
 Version 1.4.1 (released 2019-12-07)
 -----------------------------------